Squashed commit of the following:
commit df641506885e09e63f80efc6c658d01a1a3a8142
Author: James Agnew <jamesagnew@gmail.com>
Date:   Sat Sep 29 13:47:22 2018 -0400

    Shrink validation resources

commit 4c1550eafa26f4235b4b8804741aecc03276b5c6
Author: James Agnew <jamesagnew@gmail.com>
Date:   Fri Sep 28 22:45:28 2018 -0400

    Updates to get R4 working

commit 8332f15291006691ed9b07ead3e3524b0fc85510
Author: jamesagnew <jamesagnew@gmail.com>
Date:   Fri Sep 28 09:02:13 2018 -0400

    Ongoing work on gettign R4 working

commit f7146cab7aed937bb625a8aec95744d76db00041
Author: James Agnew <jamesagnew@gmail.com>
Date:   Fri Sep 28 05:21:01 2018 -0400

    More work on sync

commit f48de4a10b1bab2584fee813017d9b27b237bfa9
Merge: 9e4f3cc722 aacb78b779
Author: James Agnew <jamesagnew@gmail.com>
Date:   Thu Sep 27 20:11:29 2018 -0400

    Merge branch 'sync_r4' of github.com:jamesagnew/hapi-fhir into sync_r4

commit 9e4f3cc722668aae31a77b19bf9b032af3ae01b0
Author: James Agnew <jamesagnew@gmail.com>
Date:   Thu Sep 27 20:11:19 2018 -0400

    Fix compile error

commit 7ec29e0ceda8f9013cbabb95767cbe15b4b81303
Merge: 23f7517325 002c4b3ff7
Author: James Agnew <jamesagnew@gmail.com>
Date:   Thu Sep 27 20:04:23 2018 -0400

    Merge branch 'master' into sync_r4

commit aacb78b7793ab2c71e0a2a774240ee2b5d3f9e7f
Author: jamesagnew <jamesagnew@gmail.com>
Date:   Thu Sep 27 20:01:42 2018 -0400

    Keep working on getting R4 building

commit 4950de46d99897e74d41dfdb10d3d8a3435cb0c9
Author: jamesagnew <jamesagnew@gmail.com>
Date:   Tue Sep 25 13:11:27 2018 -0400

    Ongoing work to get R4 working

commit 819d69c20e112dd3c72569ec50114ddb8263c300
Author: jamesagnew <jamesagnew@gmail.com>
Date:   Thu Sep 20 08:15:36 2018 -0400

    Work on getting build working

commit 2c61b6cd1205ced5d7b822cf10942c69272ea078
Merge: 16b5bb06c8 dfb4de86f1
Author: jamesagnew <jamesagnew@gmail.com>
Date:   Thu Sep 20 05:16:53 2018 -0400

    Merge branch 'master' into sync_r4

commit 16b5bb06c8cdaf8d67c3b80f5aa9be9fccd99aef
Author: jamesagnew <jamesagnew@gmail.com>
Date:   Mon Sep 17 05:30:39 2018 -0400

    Work on R4 sync

commit 23f7517325a14dd2ca1eb5641296ead2776634ae
Author: James Agnew <jamesagnew@gmail.com>
Date:   Sun Sep 16 10:19:00 2018 -0400

    Work on sync

commit 6cc413c1f1dea538295aa4c16c21a5677494ac4c
Author: James Agnew <jamesagnew@gmail.com>
Date:   Fri Sep 14 17:28:28 2018 -0400

    Work on R$ sync

commit df6f6ad2ce783b07ccc383134705d874bc5d2cf1
Author: jamesagnew <jamesagnew@gmail.com>
Date:   Fri Sep 14 08:33:07 2018 -0400

    Work on R4

+++ hapi-fhir-utilities/src/main/java/org/hl7/fhir/utilities/xls/XLSXmlNormaliser.java
@@ -37,17 +37,20 @@ public class XLSXmlNormaliser {
 
   private String source;
   private String dest;
+  private boolean exceptionIfExcelNotNormalised;
   
-  public XLSXmlNormaliser(String source, String dest) {
+  public XLSXmlNormaliser(String source, String dest, boolean exceptionIfExcelNotNormalised) {
     super();
     this.source = source;
     this.dest = dest;
+    this.exceptionIfExcelNotNormalised = exceptionIfExcelNotNormalised;
   }
   
-  public XLSXmlNormaliser(String source) {
+  public XLSXmlNormaliser(String source, boolean exceptionIfExcelNotNormalised) {
     super();
     this.source = source;
     this.dest = source;
+    this.exceptionIfExcelNotNormalised = exceptionIfExcelNotNormalised;
   }
   
   public void go() throws FHIRException, TransformerException, ParserConfigurationException, SAXException, IOException {
@@ -68,6 +71,9 @@ public class XLSXmlNormaliser {
     }
     if (hasComment)
       return;
+    if (exceptionIfExcelNotNormalised)
+      throw new FHIRException("The spreadsheet "+dest+" was committed after editing in excel, but before the build could run *after Excel was closed*");
+    
     System.out.println("normalise: "+source);
     
     XMLUtil.deleteByName(root, "ActiveSheet");
@@ -82,10 +88,6 @@ public class XLSXmlNormaliser {
     
     if (!hasComment)
       root.appendChild(xml.createComment("canonicalized"));
-    String altName = dest+".please-close-this-in-excel-and-return-the-build-prior-to-committing";
-    File f = new File(altName);
-    if (f.exists())
-      f.delete();
     try {
       saveXml(new FileOutputStream(dest));
       String s = TextFile.fileToString(dest);
@@ -94,7 +96,7 @@ public class XLSXmlNormaliser {
       TextFile.stringToFile(s, dest, false);
       new File(dest).setLastModified(time);
     } catch (Exception e) {
-      TextFile.stringToFile("Run process helper", altName);
+      System.out.println("The file "+dest+" is still open in Excel, and you will have to run the build after closing Excel before committing");
     }
   }
